backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pdfplumber>=0.10.0
python-docx==0.8.11
pandas==2.1.3
scikit-learn==1.3.2
sqlalchemy==2.0.23
python-multipart==0.0.6
reportlab==4.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
python-dotenv==1.0.0
